@@ -1,29 +1,55 @@
-source :rubygems
+source 'http://rubygems.org'
 
-gem 'rails', '3.0.7'
+gem 'rails', '3.1.0'
+# locked to 1.3.3 to resolve annoying warning 'already initialized constant WFKV_'
+gem 'rack' , '1.3.3'
 
-group :test, :development do
-  gem "rspec-rails", "~> 2.5.0"
-end
-
-gem 'quality-measure-engine', '1.0.1'
+gem 'quality-measure-engine', '1.0.3'
+#gem 'quality-measure-engine', :path => '../quality-measure-engine/'
+#gem 'quality-measure-engine', :git => 'http://github.com/pophealth/quality-measure-engine.git', :branch => 'develop'
 gem 'nokogiri'
 gem 'rubyzip'
 
-gem "will_paginate", "~> 3.0.pre2"
+gem "will_paginate"
 gem 'json', :platforms => :jruby
-gem 'bson_ext', :platforms => :mri
-gem "mongoid", "2.0.1"
+# these are all tied to 1.3.1 because bson 1.4.1 was yanked.  To get bundler to be happy we need to force 1.3.1 to cause the downgrade
+gem "mongo", "1.3.1"
+gem "bson", "1.3.1"
+gem 'bson_ext',"1.3.1",  :platforms => :mri
+gem "mongoid"
 gem 'devise'
 gem 'foreman'
+gem 'pry'
+gem 'formtastic'
+gem 'cancan'
 
 # Windows doesn't have syslog, so need a gem to log to EventLog instead
 gem 'win32-eventlog', :platforms => [:mswin, :mingw]
 
-# Bundle edge Rails instead:
-# gem 'rails', :git => 'git://github.com/rails/rails.git'
+# Gems used only for assets and not required
+# in production environments by default.
+group :assets do
+  gem 'sass-rails', "  ~> 3.1.0"
+  gem 'coffee-rails', "~> 3.1.0"
+  gem 'uglifier'
+end
 
-#gem 'sqlite3-ruby', :require => 'sqlite3'
+group :test, :develop do
+  # gem "rspec-rails"
+  # Pretty printed test output
+  gem "unicorn"
+  gem 'turn', :require => false
+  gem 'cover_me'
+  gem 'factory_girl'
+  gem 'minitest'
+  gem 'mocha', :require => false
+end
+
+group :production do
+  gem 'therubyracer'
+end
+
+gem 'jquery-rails'
 
 # Use unicorn as the web server
 # gem 'unicorn'
@@ -32,17 +58,5 @@ gem 'win32-eventlog', :platforms => [:mswin, :mingw]
 # gem 'capistrano'
 
 # To use debugger
-# gem 'ruby-debug'
+# gem 'ruby-debug19', :require => 'ruby-debug'
 
-# Bundle the extra gems:
-# gem 'bj'
-# gem 'nokogiri'
-# gem 'sqlite3-ruby', :require => 'sqlite3'
-# gem 'aws-s3', :require => 'aws/s3'
-
-# Bundle gems for the local environment. Make sure to
-# put test-only gems in this group so their generators
-# and rake tasks are available in development mode:
-# group :development, :test do
-#   gem 'webrat'
-# end
