@@ -1,25 +1,31 @@
 class AdminController < ApplicationController
 
   before_filter :authenticate_user!
-  before_filter :admin_user?
+  before_filter :validate_authorization!
+  add_breadcrumb 'admin', :admin_users_path
 
   def users
-    @users = User.all
+    @users = User.all.ordered_by_username
   end
 
   def promote
-    toggle_admin_privilidges(params[:username], :promote)
+    toggle_privilidges(params[:username], params[:role], :promote)
   end
 
   def demote
-    toggle_admin_privilidges(params[:username], :demote)
+    toggle_privilidges(params[:username], params[:role], :demote)
   end
 
-  def destroy
-    user = User.first(:conditions => {:username => params[:username]})
+  def disable
+    user = User.by_username(params[:username]);
+    disabled = params[:disabled].to_i == 1
     if user
-      user.destroy
-      render :text => "removed"
+      user.update_attribute(:disabled, disabled)
+      if (disabled)
+        render :text => "<a href=\"#\" class=\"disable\" data-username=\"#{user.username}\">disabled</span>"
+      else
+        render :text => "<a href=\"#\" class=\"enable\" data-username=\"#{user.username}\">enabled</span>"
+      end
     else
       render :text => "User not found"
     end
@@ -35,25 +41,31 @@ class AdminController < ApplicationController
     end
   end
 
+  def update_npi
+    user = User.by_username(params[:username]);
+    user.update_attribute(:npi, params[:npi]);
+    render :text => "true"
+  end
+
   private
 
-  def toggle_admin_privilidges(username, direction)
-    user = User.first(:conditions => {:username => username})
+  def toggle_privilidges(username, role, direction)
+    user = User.by_username username
+
     if user
       if direction == :promote
-        user.update_attribute(:admin, true)
-        render :text => "Yes - <span class=\"demote\" data-username=\"#{username}\">remove admin rights</span>"
+        user.update_attribute(role, true)
+        render :text => "Yes - <a href=\"#\" class=\"demote\" data-role=\"#{role}\" data-username=\"#{username}\">revoke</a>"
       else
-        user.update_attribute(:admin, false)
-        render :text => "No - <span class=\"promote\" data-username=\"#{username}\">grant admin rights</span>"
+        user.update_attribute(role, false)
+        render :text => "No - <a href=\"#\" class=\"promote\" data-role=\"#{role}\" data-username=\"#{username}\">grant</a>"
       end
     else
       render :text => "User not found"
     end
   end
 
-  def admin_user?
-    redirect_to '/' unless current_user.admin?
+  def validate_authorization!
+    authorize! :admin, :users
   end
-
 end
