@@ -6,25 +6,47 @@ PopHealth::Application.routes.draw do
   post "admin/promote"
   post "admin/demote"
   post "admin/approve"
-  post "admin/destroy"
+  post "admin/disable"
+  post "admin/update_npi"
 
   get "logs/index"
-
+  
   match 'measures', :to => 'measures#index', :as => :dashboard, :via => :get
-  match 'measures/show/:id(/:sub_id)', :to => 'measures#show', :as => :measure, :via => :get
+  match "measure/:id(/:sub_id)/providers", :to => "measures#providers", :via => :get
+  match 'measure/:id(/:sub_id)', :to => 'measures#show', :as => :measure, :via => :get
   match 'measures/result/:id(/:sub_id)', :to => 'measures#result', :as => :measure_result, :via => :get
   match 'measures/definition/:id(/:sub_id)', :to => 'measures#definition', :as => :measure_definition, :via => :get
   match 'measures/patients/:id(/:sub_id)', :to => 'measures#patients', :as => :patients, :via => :get
   match 'measures/select/:id', :to => 'measures#select', :as => :select, :via => :post
-  match 'measures/remove/:id', :to => 'measures#remove', :as => :remove, :via => :post
+  match 'measure/:id/remove', :to => 'measures#remove', :as => :remove, :via => :post
   match 'measures/measure_patients/:id(/:sub_id)', :to=>'measures#measure_patients', :as => :measure_patients, :via=> :get
-  match 'measures/report', :to=>'measures#measure_report', :as => :measure_report, :via=> :get
+  match 'measures/measure_report', :to=>'measures#measure_report', :as => :measure_report, :via=> :get
   match 'measures/patient_list/:id(/:sub_id)', :to=>'measures#patient_list', :as => :patient_list, :via=> :get
   match 'measures/period', :to=>'measures#period', :as => :period, :via=> :post
-
+  
+  match 'provider/:npi', :to => "measures#index", :as => :provider_dashboard, :via => :get
+  
   match 'records', :to => 'records#create', :via => :post
+  
+  match 'patients', :to => 'patients#index', :via => :get
+  match 'patients/show/:id', :to => 'patients#show'
 
+  match 'providers/measure/:measure_id(/:sub_id)', :to => "providers#measure", :as => :providers_measure, :via => :get 
+  
   root :to => 'measures#index'
+  
+  resources :measures do
+    member { get :providers }
+  end
+  
+  resources :providers do
+    member do
+      get :merge_list
+      put :merge
+    end
+  end
+  
+  resources :teams
   
   # The priority is based upon order of creation:
   # first created -> highest priority.
@@ -75,7 +97,7 @@ PopHealth::Application.routes.draw do
 
   # You can have the root of your site routed with "root"
   # just remember to delete public/index.html.
-  # root :to => "welcome#index"
+  # root :to => 'welcome#index'
 
   # See how all your routes lay out with "rake routes"
 
